Split OpenAPI codegen sources
