+++ cli/src/main/scala/sttp/tapir/cli/Main.scala
@@ -1,0 +1,3 @@
+package sttp.tapir.cli
+
+object Main
