--- sbt/sbt-openapi-codegen/src/main/scala/codegen/YamlParser.scala
@@ -1,16 +1,0 @@
-package codegen
-
-import codegen.openapi.models.OpenapiModels.OpenapiDocument
-import io.circe.yaml.parser
-
-object YamlParser {
-  import cats.implicits._
-  import io.circe._
-
-  def parseFile(yamlString: String): Either[Error, OpenapiDocument] = {
-    parser
-      .parse(yamlString)
-      .leftMap(err => err: Error)
-      .flatMap(_.as[OpenapiDocument])
-  }
-}
