--- sbt/sbt-openapi-codegen/src/main/scala/codegen/openapi/models/OpenapiComponent.scala
@@ -1,15 +1,0 @@
-package codegen.openapi.models
-
-case class OpenapiComponent(schemas: Map[String, OpenapiSchemaType])
-
-object OpenapiComponent {
-  import io.circe._
-
-  implicit val OpenapiComponentDecoder: Decoder[OpenapiComponent] = { (c: HCursor) =>
-    for {
-      schemas <- c.downField("schemas").as[Option[Map[String, OpenapiSchemaType]]]
-    } yield {
-      OpenapiComponent(schemas.getOrElse(Map.empty))
-    }
-  }
-}
