--- sbt/sbt-openapi-codegen/src/sbt-test/sbt-openapi-codegen/caching/project/build.properties
@@ -1,1 +1,0 @@
-sbt.version=1.3.13
