--- sbt/sbt-openapi-codegen/src/sbt-test/sbt-openapi-codegen/caching/project/plugins.sbt
@@ -1,11 +1,0 @@
-{
-  val pluginVersion = System.getProperty("plugin.version")
-  if (pluginVersion == null)
-    throw new RuntimeException("""|
-                                  |
-                                  |The system property 'plugin.version' is not defined.
-                                  |Specify this property using the scriptedLaunchOpts -D.
-                                  |
-                                  |""".stripMargin)
-  else addSbtPlugin("com.softwaremill.sttp.tapir" % "sbt-openapi-codegen" % pluginVersion)
-}
