--- sbt/sbt-openapi-codegen/src/sbt-test/sbt-openapi-codegen/caching/test
@@ -1,6 +1,0 @@
-> clean
-> run
-> check
-$ copy-file target/scala-2.12/src_managed/main/sbt-openapi-codegen/TapirGeneratedEndpoints.scala target/TapirGeneratedEndpoints.scala
-> compile
-$ newer target/TapirGeneratedEndpoints.scala target/scala-2.12/src_managed/main/sbt-openapi-codegen/TapirGeneratedEndpoints.scala
