--- sbt/sbt-openapi-codegen/src/sbt-test/sbt-openapi-codegen/minimal/test
@@ -1,5 +1,0 @@
-> clean
-> generateTapirDefinitions
-> run
-> check
-
