--- sbt/sbt-openapi-codegen/src/sbt-test/sbt-openapi-codegen/option-overrides/example_swagger.yaml
@@ -1,54 +1,0 @@
-openapi: 3.0.1
-info:
-  title: My Bookshop
-  version: '1.0'
-paths:
-  /books/{genre}/{year}:
-    get:
-      operationId: getBooksGenreYear
-      parameters:
-      - name: genre
-        in: path
-        required: true
-        schema:
-          type: string
-      - name: year
-        in: path
-        required: true
-        schema:
-          type: integer
-      - name: limit
-        in: query
-        description: Maximum number of books to retrieve
-        required: true
-        schema:
-          type: integer
-      - name: X-Auth-Token
-        in: header
-        required: true
-        schema:
-          type: string
-      responses:
-        '200':
-          description: ''
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/Book'
-        default:
-          description: ''
-          content:
-            text/plain:
-              schema:
-                type: string
-components:
-  schemas:
-    Book:
-      required:
-      - title
-      type: object
-      properties:
-        title:
-          type: string
