--- sbt/sbt-openapi-codegen/src/sbt-test/sbt-openapi-codegen/option-overrides/test
@@ -1,5 +1,0 @@
-> clean
-> generateTapirDefinitions
-> run
-> check
-
