--- sbt/sbt-openapi-codegen/src/test/scala/codegen/BasicGeneratorSpec.scala
@@ -1,11 +1,0 @@
-package codegen
-
-import codegen.testutils.CompileCheckTestBase
-
-class BasicGeneratorSpec extends CompileCheckTestBase {
-
-  it should "generate the bookshop example" in {
-    BasicGenerator.generateObjects(TestHelpers.myBookshopDoc, "sttp.tapir.generated", "TapirGeneratedEndpoints") shouldCompile ()
-  }
-
-}
